Creating @airswap/metadata with initial rinkeby tokens
